modular components with compound pattern
