+++ src/components/sheets/local-components/Editors/Elements/CodeElement.jsx
@@ -1,0 +1,9 @@
+const CodeElement = props => {
+    return (
+      <pre {...props.attributes} className='codestyle'>
+        <code>{props.children}</code>
+      </pre>
+    )
+  }
+  
+  export default CodeElement
