+++ src/components/sheets/local-components/Editors/Elements/DefaultElement.jsx
@@ -1,0 +1,5 @@
+const DefaultElement = props => {
+    return <p {...props.attributes}>{props.children}</p>
+  }
+
+  export default DefaultElement
