+++ src/components/sheets/local-components/Editors/leaves/Leaf.jsx
@@ -1,0 +1,12 @@
+const Leaf = props => {
+  return (
+    <span
+      {...props.attributes}
+      style={{ fontWeight: props.leaf.bold ? 'bold' : 'normal' }}
+    >
+      {props.children}
+    </span>
+  )
+}
+
+export default Leaf
